// src/use/controller/jobs/jobs.d.ts
export interface IJobRecord {
  _date: number;
  id: string;
  name: string;
  note: string;
  companyid: string;
  contactid: string;
  owner: string;
  frequency: number;
  duedate: number;
  duration: number;
  actualduration: number;
  rejected: number;
  checkoutnote: string;
  status: number;
  street: string;
  zipcode: string;
  city: string;
  state: string;
  country: string;
  geo: string;
  groupid: string;
  assetid: string;
  siteid: string;
  formid: string;
  priority: number;
  signature: string;
  custom: string;
  waypoint: number;
  recurringid: string;
  quoteid: string;
  invoiceid: string;
  projectid: string;
  expirydate: number;
  creationdate: number;
  enddate: number;
  toolid: string;
  history: string;
  checkout: number;
  checkoutgeo: string;
  mandatorytemplateid: string;
  counterid: string;
  changelog: string;
  companyName?: string;
  contactName?: string;
  groupName?: string;
  assetName?: string;
  formName?: string;
  projectName?: string;
  toolName?: string;
  quoteName?: string;
  statusName?: string;
  CustomerPhoneNumber?: string | null;
  JobPhase?: string | null;
  ActivityReportSent?: string | null;
  HoursType?: string | null;
  CreatedinException_Finalreport?: string | null;
  RLD?: string | null;
  UpdateReason?: string | null;
  RetrieverStatus?: string | null;
  SLAPriority?: string | null;
  Updated?: string | null;
  ProposedEnd?: string | null;
  ProposedStart?: string | null;
  JobType?: string | null;
  Alert?: string | null;
  FormID?: string | null;
  PlantSection?: string | null;
  MapSection?: string | null;
  "NotificationCode:"?: string | null;
  "NotificationType:"?: string | null;
  "NotificationNumber:"?: string | null;
  AusgridPoleInspection?: string | null;
  Description?: string | null;
  Prioritytext?: string | null;
  RequiredStart?: string | null;
  RequiredEnd?: string | null;
  Coding?: string | null;
  Userstatus?: string | null;
  Systemstatus?: string | null;
  PoleId?: string | null;
  PoleGrid?: string | null;
  Notification?: string | null;
  NotificationType?: string | null;
  Vector?: string | null;
  Pole?: string | null;
  Subject?: string | null;
  UTCSubReason?: string | null;
  SpecialInstructions?: string | null;
  FSPWorkflowstatus?: string | null;
  "AppointmentEndtime:"?: number | null;
  "AppointmentStarttime:"?: number | null;
  NMINumber?: number | null;
}